c1

+++ pybricks/pyBricks1/shared.py
@@ -8,7 +8,7 @@ class Speed:
     Slow = float(150)
     Medium = float(300)
     # 600 is max, but sometimes gyro make mistakes so, I guess 500 is sweet spot
-    Fast = float(500)
+    Fast = float(450)
 
 
 # inspiration https://github.com/FLL-Team-24277/FLL-Fall-2023-Masterpiece/blob/main/samples/base_robot.py
